+++ src/main/java/net/pms/newgui/GuiUtil.java
@@ -96,7 +96,7 @@ public final class GuiUtil {
 			int filled = getAmountFull(b, w, h);
 			Graphics2D g2 = (Graphics2D)g;
 			g2.setColor(progressBar.getForeground());
-			g2.setStroke(new BasicStroke((float)h, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL));
+			g2.setStroke(new BasicStroke(h, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL));
 			g2.drawLine(b.left, (h/2) + b.top, filled + b.left, (h/2) + b.top);
 			// Draw the string, if any
 			if (progressBar.isStringPainted()) {
@@ -315,7 +315,7 @@ public final class GuiUtil {
 				g2.setColor(Color.gray);
 				paintTicks(g, b.left, (int)(unit * tickmarks), b.left + w);
 			}
-			g2.setStroke(new BasicStroke((float) h, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL));
+			g2.setStroke(new BasicStroke(h, BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL));
 			// Draw the segments
 			for (Segment s : segments) {
 				if (s.val > 0) {
@@ -369,7 +369,7 @@ public final class GuiUtil {
 			// Draw the max value if it's not at a tick mark and we have room
 			if (tickLabel != null && ((max - x0) % step != 0)) {
 				int avail_w = max - x + step;
-				String s = "" + (int)progressBar.getMaximum();
+				String s = "" + progressBar.getMaximum();
 				int w = (int)g.getFontMetrics().getStringBounds(s, g).getWidth();
 				if (avail_w > w + 10) {
 					g.drawString(s, max - 3 - w, h - 3);
